lights code update
