+++ lights.py
@@ -1,10 +1,13 @@
 import pigpio
 import time
 from logpy.LogPy import Logger
+from ports import LIGHTS_DRIVER_PORT
+from rov_comm import Client
 
 PWM_lights = 27
 ON_FULL = 600
 OFF = 1400
+
 
 class LightHandling:
     def __init__(self):
@@ -29,6 +32,18 @@ class LightHandling:
         x=100-x
         self.pi.set_PWM_dutycycle(PWM_lights, -8*x+1400)
         self.logger.log("lights' brightness: " + str(x) + "%")
+
+class Lights:
+
+    def __init__(self):
+            self.client_brightness = Client(LIGHTS_DRIVER_PORT)
+            self.light_handling = LightHandling()
+
+    def set_brightness(self):
+        command = self.client_brightness.get_data()
+        #print(command) 
+        x = int(command['power'])
+        self.light_handling.lights_set_brightness(x)
 
 if __name__ == "__main__":
     LIGHTS = LightHandling()
